ex3 - filter duplicates
For each author node in the graph, DFS is run to retrieve the distance
from the author to each coauthor. Cycles are avoided by using a sequence
(used as a stack). But even if cycles are avoided, duplicates are still
possible, because there are many paths that can go from the author to
one coauthor.

Each time DFS is run on one node, duplicates are filtered for this node by
taking the distance entry which is minimal.